+++ homeassistant/components/soma/__init__.py
@@ -1,9 +1,7 @@
 """Support for Soma Smartshades."""
 import asyncio
-import logging
 
 from api.soma_api import SomaApi
-from requests import RequestException
 import voluptuous as vol
 
 from homeassistant import config_entries
@@ -16,8 +14,6 @@ from homeassistant.helpers.typing import HomeAssistantType
 from .const import API, DOMAIN, HOST, PORT
 
 DEVICES = "devices"
-
-_LOGGER = logging.getLogger(__name__)
 
 CONFIG_SCHEMA = vol.Schema(
     {
@@ -113,43 +109,3 @@ class SomaEntity(Entity):
             "name": self.name,
             "manufacturer": "Wazombi Labs",
         }
-
-    async def async_update(self):
-        """Update the device with the latest data."""
-        try:
-            response = await self.hass.async_add_executor_job(
-                self.api.get_shade_state, self.device["mac"]
-            )
-        except RequestException:
-            _LOGGER.error("Connection to SOMA Connect failed")
-            self.is_available = False
-            return
-        if response["result"] != "success":
-            _LOGGER.error(
-                "Unable to reach device %s (%s)", self.device["name"], response["msg"]
-            )
-            self.is_available = False
-            return
-        self.current_position = 100 - response["position"]
-        try:
-            response = await self.hass.async_add_executor_job(
-                self.api.get_battery_level, self.device["mac"]
-            )
-        except RequestException:
-            _LOGGER.error("Connection to SOMA Connect failed")
-            self.is_available = False
-            return
-        if response["result"] != "success":
-            _LOGGER.error(
-                "Unable to reach device %s (%s)", self.device["name"], response["msg"]
-            )
-            self.is_available = False
-            return
-        # https://support.somasmarthome.com/hc/en-us/articles/360026064234-HTTP-API
-        # battery_level response is expected to be min = 360, max 410 for
-        # 0-100% levels above 410 are consider 100% and below 360, 0% as the
-        # device considers 360 the minimum to move the motor.
-        _battery = round(2 * (response["battery_level"] - 360))
-        battery = max(min(100, _battery), 0)
-        self.battery_state = battery
-        self.is_available = True
